Replace custom retry code with retry mechanism of Faraday
